SDK_LEARNING/LEARNING_WITH_ONLINE_CLASS/PRACTICE_CODE/TRAICING/pyproject.toml
[project]
name = "tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai[standard]>=1.86.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
